+++ README.md
@@ -23,7 +23,7 @@ r.Get("/todos", quickapi.Lift(ListTodo))
 
 ## experimental openapi support
 
-:warning: this is experimental feature, using quickapi/experimental/define package can be able to define openapi doc with define router.
+:warning: this is experimental feature, using quickapi/qopenapi/define package can be able to define openapi doc with define router.
 
 
 ## how to use
