+++ qopenapi/define/skeleton.json
