+++ qopenapi/define/type.go
